+++ Forge.Forms/src/Forge.Forms/Annotations/TimeAttribute.cs
@@ -5,5 +5,10 @@ namespace Forge.Forms.Annotations
     [AttributeUsage(AttributeTargets.Property)]
     public sealed class TimeAttribute : Attribute
     {
+        /// <summary>
+        /// Determines whether the time is displayed in 24-hour format.
+        /// Accepts a boolean or a dynamic resource.
+        /// </summary>
+        public object Is24Hours { get; set; } = false;
     }
 }