+++ Forge.Forms/src/Forge.Forms/FormBuilding/Defaults/TimeField.cs
@@ -1,23 +1,28 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Windows;
+using Forge.Forms.DynamicExpressions;
 
 namespace Forge.Forms.FormBuilding.Defaults
 {
     public class TimeField : DataFormField
     {
-        /// <summary>
-        /// </summary>
-        /// <param name="key"></param>
         public TimeField(string key) : base(key, typeof(DateTime))
         {
+        }
+
+        public IValueProvider Is24Hours { get; set; }
+
+        protected internal override void Freeze()
+        {
+            base.Freeze();
+            Resources.Add(nameof(Is24Hours), Is24Hours ?? LiteralValue.False);
         }
 
         protected internal override IBindingProvider CreateBindingProvider(IResourceContext context,
             IDictionary<string, IValueProvider> formResources)
         {
-            var datePresenter = new TimePresenter(context, Resources, formResources);
-            return datePresenter;
+            return new TimePresenter(context, Resources, formResources);
         }
     }